+++ src/widgets/checkbox.rs
@@ -70,6 +70,7 @@ impl<'a> StyledCheckbox<'a> {
                 })
                 .inner;
 
+            SharedStyle::paint_widget_side_borders(ui, &response, &per);
             paint_shadows(
                 ui,
                 shadow_idx,